Moved all files out of "static" directory
Files were previously in a “static” directory because the local player
was copied from the static files in the server application.  Since the
local player is meant to stand alone, there is no longer any need for
them to be in a “static” directory.
- also removed “FileSaver.js” (recommend getting the current version
from https://github.com/eligrey/FileSaver.js/ instead)
- removed a file that was not intended to be included